Initial commit
This started as test to see what it would look like if I replicated one
of the first video games I had owned as a child. The test was successful
and I decided to pursue it to the end.

This first commit has a console and a view that replicates the actual
game play. Nothing is hooked up but you can change the direction and
location manually in the code to look around the dungeon.